requirements.txt
python-telegram-bot==21.9
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
python-dotenv==1.0.1
greenlet==3.1.1
tzdata>=2024.1
